@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 
 import { FC, useCallback, useState, useEffect } from "react";
-import { createStyles, IconButton, Theme, makeStyles } from "@material-ui/core";
-import BookmarkIcon from "@material-ui/icons/Bookmark";
-import PropTypes from "prop-types";
-import { orange, grey } from "@material-ui/core/colors";
+import { makeStyles } from "@mui/styles";
+import { createStyles, IconButton, Theme } from "@mui/material";
+import BookmarkIcon from "@mui/icons-material/Bookmark";
+import { orange, grey } from "@mui/material/colors";
 import ShowHideGuard from "src/components/auth/guards/ShowHideGuard";
 import { apiService } from "src/utils/apiService";
 // import useAuth from "src/hooks/useAuth";
@@ -56,10 +56,6 @@ const Bookmark: FC<any> = ({ pageId }) => {
       </IconButton>
     </ShowHideGuard>
   );
-};
-
-Bookmark.propTypes = {
-  pageId: PropTypes.any,
 };
 
 export default Bookmark;
