@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import type { FC } from "react";
 import NProgress from "nprogress";
-import { Box } from "@material-ui/core";
+import { Box } from "@mui/material";
 
 const LoadingScreen: FC = () => {
   useEffect(() => {
