@@ -1,5 +1,5 @@
 import type { FC } from "react";
-import { Box } from "@material-ui/core";
+import { Box } from "@mui/material";
 import Logo from "./Logo";
 
 const SlashScreen: FC = () => (
