@@ -1,10 +1,8 @@
 import type { FC, ReactNode } from "react";
-import PropTypes from "prop-types";
 // import { matchPath } from 'react-router-dom';
 import { matchPath } from "src/utils/matchPath";
 
-import { List, ListSubheader } from "@material-ui/core";
-import type { ListProps } from "@material-ui/core";
+import { List, ListSubheader, ListProps } from "@mui/material";
 import NavItem from "./NavItem";
 
 interface Item {
@@ -131,12 +129,6 @@ const NavSection: FC<NavSectionProps> = (props) => {
       })}
     </List>
   );
-};
-
-NavSection.propTypes = {
-  items: PropTypes.array,
-  pathname: PropTypes.string,
-  title: PropTypes.string,
 };
 
 export default NavSection;
