@@ -1,8 +1,13 @@
 import type { FC } from "react";
 import { Helmet } from "react-helmet-async";
-import { Box, Button, Container, Typography } from "@material-ui/core";
-import useMediaQuery from "@material-ui/core/useMediaQuery";
-import { useTheme } from "@material-ui/core/styles";
+import {
+  Box,
+  Button,
+  Container,
+  Typography,
+  useMediaQuery,
+  useTheme,
+} from "@mui/material";
 import { useRouter } from "next/router";
 
 const NotFound: FC = () => {
