@@ -1,10 +1,8 @@
 import { useState } from "react";
 import type { FC, ReactNode } from "react";
-import PropTypes from "prop-types";
-import { experimentalStyled } from "@material-ui/core/styles";
+import { experimentalStyled } from "@mui/material";
 import Footer from "../Footer";
 import MainNavbar from "./MainNavbar";
-import MainSidebar from "./MainSidebar";
 
 interface MainLayoutProps {
   children?: ReactNode;
@@ -17,26 +15,13 @@ const MainLayoutRoot = experimentalStyled("div")(({ theme }) => ({
 }));
 
 const MainLayout: FC<MainLayoutProps> = ({ children }) => {
-  const [isSidebarMobileOpen, setIsSidebarMobileOpen] =
-    useState<boolean>(false);
-
   return (
     <MainLayoutRoot>
-      <MainNavbar
-        onSidebarMobileOpen={(): void => setIsSidebarMobileOpen(true)}
-      />
-      <MainSidebar
-        onMobileClose={(): void => setIsSidebarMobileOpen(false)}
-        openMobile={isSidebarMobileOpen}
-      />
+      <MainNavbar />
       {children}
       <Footer />
     </MainLayoutRoot>
   );
-};
-
-MainLayout.propTypes = {
-  children: PropTypes.node,
 };
 
 export default MainLayout;
