@@ -1,22 +1,13 @@
 import React, { FC } from "react";
-import Card from "@material-ui/core/Card";
-import CardActionArea from "@material-ui/core/CardActionArea";
-import CardActions from "@material-ui/core/CardActions";
-import CardContent from "@material-ui/core/CardContent";
-import CardMedia from "@material-ui/core/CardMedia";
-import Button from "@material-ui/core/Button";
-import Typography from "@material-ui/core/Typography";
 import {
-  createStyles,
-  IconButton,
-  Theme,
-  makeStyles,
-  Grid,
-  Paper,
-} from "@material-ui/core";
-import Bookmark from "../Bookmark";
+  Card,
+  CardActionArea,
+  CardContent,
+  CardMedia,
+  Typography,
+} from "@mui/material";
+import { makeStyles } from "@mui/styles";
 import Link from "next/link";
-import { red } from "@material-ui/core/colors";
 
 const useStyles = makeStyles({
   root: {
