@@ -1,8 +1,6 @@
-/* eslint-disable no-new */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import { useRef, useState, ReactElement } from "react";
+import { useState, ReactElement } from "react";
 import { Helmet } from "react-helmet-async";
-import { Box, Card, CardContent, CardHeader, Grid } from "@material-ui/core";
+import { Grid, Box, Card, CardContent } from "@mui/material";
 import DashboardLayout from "src/components/dashboard-layout/DashboardLayout";
 import axios from "axios";
 import FirstForm from "src/components/timeseries-sim/FirstForm";
@@ -47,7 +45,6 @@ const initialFirstRunState = {
 const TimeseriesSim = () => {
   //const { settings } = useSettings();
   const classes = useStyles();
-  const ref = useRef();
 
   const [advancedFilterPopup, setAdvancedFilterPopup] = useState(false);
   const advancedFilterClickOpen = () => {
@@ -130,7 +127,6 @@ const TimeseriesSim = () => {
           </Grid>
           <Grid>
             <Card sx={{ width: 1 }}>
-              <CardHeader />
               <CardContent>
                 <div id="alignmentchart" />
               </CardContent>
