@@ -1,8 +1,6 @@
 import type { FC } from "react";
-import PropTypes from "prop-types";
-import { AppBar, Box, IconButton, Toolbar } from "@mui/material";
+import { AppBar, Box, IconButton, Toolbar, AppBarProps } from "@mui/material";
 import { experimentalStyled } from "@mui/material/styles";
-import type { AppBarProps } from "@mui/material";
 import MenuIcon from "@mui/icons-material//Menu";
 import AccountPopover from "./AccountPopover";
 
@@ -21,7 +19,6 @@ const DashboardNavbarRoot = experimentalStyled(AppBar)(({ theme }) => ({
     borderBottom: `1px solid ${theme.palette.divider}`,
     boxShadow: "none",
   }),
-  zIndex: theme.zIndex.drawer + 100,
 }));
 
 const DashboardNavbar: FC<DashboardNavbarProps> = (props) => {
@@ -35,7 +32,7 @@ const DashboardNavbar: FC<DashboardNavbarProps> = (props) => {
           onClick={onSidebarMobileOpen}
           sx={{
             display: {
-              lg: "none",
+              md: "none",
             },
           }}
         >
@@ -80,10 +77,6 @@ const DashboardNavbar: FC<DashboardNavbarProps> = (props) => {
       </Toolbar>
     </DashboardNavbarRoot>
   );
-};
-
-DashboardNavbar.propTypes = {
-  onSidebarMobileOpen: PropTypes.func,
 };
 
 export default DashboardNavbar;
