@@ -14,7 +14,8 @@
 
 import * as d3 from "d3";
 import { Data } from "../data.js";
-import { dashboard, COLOR_VACCINATON, COLOR_DEATHS } from "./dashboard";
+import { dashboard } from "./dashboard";
+import { colors } from "../colors.js";
 
 export class CouncilOverview {
   CHART_WIDTH = 1000;
@@ -26,25 +27,19 @@ export class CouncilOverview {
       .select("#" + options.chartElement)
       .append("div")
       .attr("class", "vis-example-container");
-    // .style('width', this.CHART_WIDTH + 'px')
-    // .style('height', this.CHART_HEIGHT + 'px')
-    // .text('an awesome visualisation');
 
-    // This could be a way to get data without relying on stream order.
-    const allDeathData = Data.from(
-      options.data,
-      Data.Fields.COUNCIL_ALL_DEATHS,
-    );
-    console.log(allDeathData);
+    var council = options.data[0].description;
+
     const covidDeathData = Data.from(
       options.data,
       Data.Fields.COUNCIL_COVID_DEATHS,
     );
+    console.log("allDeathData:", covidDeathData[0]);
+
     const vaccineData = Data.from(
       options.data,
       Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
     );
-    console.log(vaccineData);
 
     var config = {
       layout: [
@@ -56,7 +51,7 @@ export class CouncilOverview {
         {
           id: "council",
           title: "Deaths",
-          layout: [["covidDeathData", "allDeathData"]],
+          layout: ["covidDeathData"],
         },
         {
           id: "vaccinations-total",
@@ -102,20 +97,22 @@ export class CouncilOverview {
         {
           id: "covidDeathData",
           title: "Covid Deaths",
-          dataField: "Aberdeen City",
+          dataField: council,
           visualization: "linechart",
-          color: COLOR_DEATHS,
+          color: colors.getDeathColor(),
           data: Data.from(options.data, Data.Fields.COUNCIL_COVID_DEATHS),
           mode: dashboard.MODE_DAILY,
+          detail: dashboard.DETAIL_HIGHT,
           conditions: ["index.length > 4"],
         },
         {
           id: "allDeathData",
           title: "All Deaths",
-          dataField: "Aberdeen City",
+          dataField: council,
           visualization: "linechart",
-          color: d3.color(COLOR_DEATHS).brighter(0.8),
+          color: colors.getDeathColor(),
           data: Data.from(options.data, Data.Fields.COUNCIL_ALL_DEATHS),
+          detail: dashboard.DETAIL_HIGH,
           mode: dashboard.MODE_DAILY,
           conditions: ["index.length > 4"],
         },
@@ -124,12 +121,13 @@ export class CouncilOverview {
           title: "Vaccination over 18 years (1st Dose)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: COLOR_VACCINATON,
+          color: colors.getVaccinationColor(1),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
+          detail: dashboard.DETAIL_HIGH,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 1"',
@@ -141,12 +139,13 @@ export class CouncilOverview {
           title: "Vaccination over 18 years (2nd Dose)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(0.5),
+          color: colors.getVaccinationColor(2),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
+          detail: dashboard.DETAIL_HIGH,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 2"',
@@ -159,13 +158,13 @@ export class CouncilOverview {
           title: "Vaccination (18-29 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(1),
+          color: colors.getVaccinationColor(1),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 1"',
@@ -177,13 +176,13 @@ export class CouncilOverview {
           title: "Vaccination (30-39 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(1.5),
+          color: colors.getVaccinationColor(),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 1"',
@@ -195,13 +194,13 @@ export class CouncilOverview {
           title: "Vaccination (40-49 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(2),
+          color: colors.getVaccinationColor(),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 1"',
@@ -213,13 +212,13 @@ export class CouncilOverview {
           title: "Vaccination (50-54 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(2.5),
+          color: colors.getVaccinationColor(),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 1"',
@@ -231,13 +230,13 @@ export class CouncilOverview {
           title: "Vaccination (60-64 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(3),
+          color: colors.getVaccinationColor(),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 1"',
@@ -249,13 +248,13 @@ export class CouncilOverview {
           title: "Vaccination (65-69 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(3.5),
+          color: colors.getVaccinationColor(),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 1"',
@@ -267,13 +266,13 @@ export class CouncilOverview {
           title: "Vaccination (70-74 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(4),
+          color: colors.getVaccinationColor(),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 1"',
@@ -285,13 +284,13 @@ export class CouncilOverview {
           title: "Vaccination (75-79 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(4.5),
+          color: colors.getVaccinationColor(),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_CUMULATIVE,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 1"',
@@ -303,13 +302,13 @@ export class CouncilOverview {
           title: "Vaccination (80+ years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(5),
+          color: colors.getVaccinationColor(),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 1"',
@@ -322,12 +321,12 @@ export class CouncilOverview {
           title: "Vaccination (18-29 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(1),
+          color: colors.getVaccinationColor(2),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           mode: dashboard.MODE_PERCENT,
           conditions: [
             'Sex == "Total"',
@@ -340,13 +339,13 @@ export class CouncilOverview {
           title: "Vaccination (30-39 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(1.5),
+          color: colors.getVaccinationColor(2),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 2"',
@@ -358,13 +357,13 @@ export class CouncilOverview {
           title: "Vaccination (40-49 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(2),
+          color: colors.getVaccinationColor(2),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 2"',
@@ -376,13 +375,13 @@ export class CouncilOverview {
           title: "Vaccination (50-54 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(2.5),
+          color: colors.getVaccinationColor(2),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 2"',
@@ -394,13 +393,13 @@ export class CouncilOverview {
           title: "Vaccination (60-64 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(3),
+          color: colors.getVaccinationColor(2),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 2"',
@@ -412,13 +411,13 @@ export class CouncilOverview {
           title: "Vaccination (65-69 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(3.5),
+          color: colors.getVaccinationColor(2),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 2"',
@@ -430,13 +429,13 @@ export class CouncilOverview {
           title: "Vaccination (70-74 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(4),
+          color: colors.getVaccinationColor(2),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 2"',
@@ -448,13 +447,13 @@ export class CouncilOverview {
           title: "Vaccination (75-79 years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(4.5),
+          color: colors.getVaccinationColor(2),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           onditions: [
             'Sex == "Total"',
             'Dose == "Dose 2"',
@@ -466,13 +465,13 @@ export class CouncilOverview {
           title: "Vaccination (80+ years)",
           dataField: "CumulativePercentCoverage",
           visualization: "linechart",
-          color: d3.color(COLOR_VACCINATON).darker(5),
+          color: colors.getVaccinationColor(2),
           data: Data.from(
             options.data,
             Data.Fields.COUNCIL_VACCINE_SEX_AGEGROUP,
           ),
           mode: dashboard.MODE_PERCENT,
-          detail: dashboard.DETAIL_NARROW,
+          detail: dashboard.DETAIL_LOW,
           conditions: [
             'Sex == "Total"',
             'Dose == "Dose 2"',
