@@ -8,6 +8,7 @@ import { StackedAreaChart } from "./stacked-area-chart";
 import { StackedBarChart } from "./stacked-bar-chart";
 import { MirroredStackedAreaChart } from "./mirrored-stacked-area-chart";
 import { MirroredStackedBarChart } from "./mirrored-stacked-bar-chart";
+import { SensitivityStackedBarChart } from "./sensitivity-stacked-bar-chart";
 import { RiskMonitoring } from "./risk-monitoring";
 
 // Dashboards
@@ -15,6 +16,12 @@ import { CouncilOverview } from "./dashboards/council-overview";
 import { CountryOverview } from "./dashboards/country-overview";
 import { HealthBoardOverview } from "./dashboards/health-board-overview";
 import { CountryOverviewNew } from "./dashboards/country-overview-new";
+import { VaccineOverview } from "./dashboards/vaccineOverview";
+import { DashboardUK } from "./dashboards/dashboard-UK";
+import { DashboardTian } from "./dashboards/dashboard-tian";
+import { DashboardLowerTierLocalAuthority } from "./dashboards/dashboard-lowerTierLocalAuthority";
+import { DashboardMSOA } from "./dashboards/dashboard-msoa";
+import { DashboardRegion } from "./dashboards/dashboard-region"
 
 export const visFactory = (type, args) => {
   if (type === "SimpleBarChart") return new SimpleBarChart(args);
@@ -31,12 +38,21 @@ export const visFactory = (type, args) => {
     return new MirroredStackedAreaChart(args);
   if (type === "MirroredStackedBarChart")
     return new MirroredStackedBarChart(args);
+  if (type === "SensitivityStackedBarChart")
+    return new SensitivityStackedBarChart(args);
   if (type === "RiskMonitoring") return new RiskMonitoring(args);
 
   // Dashboards
   if (type === "CouncilOverview") return new CouncilOverview(args);
   if (type === "HealthBoardOverview") return new HealthBoardOverview(args);
   if (type === "CountryOverview") return new CountryOverview(args);
-  if (type === "CountryOverviewNew") return new CountryOverviewNew(args);
+  // if (type === "CountryOverviewNew") return new CountryOverviewNew(args);
+  if (type === "CountryOverviewNew") return new DashboardTian(args);
+  // if (type === "VaccineOverview") return new VaccineOverview(args);
+  if (type === "VaccineOverview") return new DashboardUK(args);
+  // if (type === "VaccineOverview") return new DashboardLowerTierLocalAuthority(args);
+  // if (type === "VaccineOverview") return new DashboardMSOA(args);
+  // if (type === "VaccineOverview") return new DashboardRegion(args);
+
   return null;
 };
