@@ -1,6 +1,7 @@
 import React from "react";
 import Document, { Html, Main, Head, NextScript } from "next/document";
-import { ServerStyleSheets } from "@material-ui/styles";
+import createEmotionServer from "@emotion/server/create-instance";
+import { createEmotionCache } from "../mui/createEmotionCache";
 
 class MyDocument extends Document {
   render() {
@@ -63,26 +64,39 @@ MyDocument.getInitialProps = async (ctx) => {
   // 3. app.render
   // 4. page.render
 
-  // Render app and page and get the context of the page with collected side effects.
-  const sheets = new ServerStyleSheets();
   const originalRenderPage = ctx.renderPage;
+
+  // You can consider sharing the same emotion cache between all the SSR requests to speed up performance.
+  // However, be aware that it can have global side effects.
+  const cache = createEmotionCache();
+  const { extractCriticalToChunks } = createEmotionServer(cache);
 
   ctx.renderPage = () =>
     originalRenderPage({
-      // eslint-disable-next-line react/display-name -- code borrowed from as is from https://github.com/mui-org/material-ui/blob/e544918cf772b4b4f715a671b4d171ad98af3663/examples/nextjs/pages/_document.js#L58
-      enhanceApp: (App) => (props) => sheets.collect(<App {...props} />),
+      // eslint-disable-next-line react/display-name,@typescript-eslint/no-explicit-any -- code borrowed from as is from https://github.com/mui-org/material-ui/blob/512896973499adbbda057e7f3685d1b23cc02de9/examples/nextjs-with-typescript/pages/_document.tsx
+      enhanceApp: (App: any) => (props) =>
+        <App emotionCache={cache} {...props} />,
     });
 
   const initialProps = await Document.getInitialProps(ctx);
+  // This is important. It prevents emotion to render invalid HTML.
+  // See https://github.com/mui-org/material-ui/issues/26561#issuecomment-855286153
+  const emotionStyles = extractCriticalToChunks(initialProps.html);
+  const emotionStyleTags = emotionStyles.styles.map((style) => (
+    <style
+      data-emotion={`${style.key} ${style.ids.join(" ")}`}
+      key={style.key}
+      // eslint-disable-next-line react/no-danger
+      dangerouslySetInnerHTML={{ __html: style.css }}
+    />
+  ));
 
   return {
     ...initialProps,
     // Styles fragment is rendered after the app and page rendering finish.
     styles: [
-      <React.Fragment key="styles">
-        {initialProps.styles}
-        {sheets.getStyleElement()}
-      </React.Fragment>,
+      ...React.Children.toArray(initialProps.styles),
+      ...emotionStyleTags,
     ],
   };
 };
